+++ src/lib.rs
@@ -17,6 +17,7 @@
 //!   * Parity checks
 //!   * Flow control
 //!   * Read/write timeouts
+//! * Full access to platform specific serial port settings using target specific feature flags (`"unix"` or `"windows"`).
 //!
 //! You can open and configure a serial port in one go with [`SerialPort::open()`].
 //! The second argument to `open()` must be a type that implements [`IntoSettings`].
@@ -48,6 +49,8 @@
 //! # }
 //! ```
 
+#![cfg_attr(feature = "doc-cfg", feature(doc_cfg))]
+
 #![warn(missing_docs)]
 
 mod sys;
@@ -60,3 +63,5 @@ pub use serial_port::SerialPort;
 
 mod settings;
 pub use settings::{CharSize, FlowControl, Parity, Settings, StopBits, COMMON_BAUD_RATES};
+
+pub mod os;